+++ src/Variants/EnumEquatableValidator.php
@@ -10,28 +10,29 @@ declare(strict_types=1);
 namespace SimpleComplex\Validate\Variants;
 
 use SimpleComplex\Validate\Type;
-use SimpleComplex\Validate\RecursiveValidator;
+use SimpleComplex\Validate\UncheckedValidator;
 
 /**
- * Unchecked validator with enum() accepting bool|int|string|null.
+ * Unchecked validator with enum() accepting bool|int|string.
  *
- * Ruleset generator will deny float bucket in allowed-values argument for enum.
- * @see Type::EQUATABLE_NULL
+ * Ruleset generator will deny float|null bucket in allowed-values argument
+ * for enum.
+ * @see Type::EQUATABLE
  *
  * @package SimpleComplex\Validate
  */
-class EnumEquatableNullRecursiveValidator extends RecursiveValidator
+class EnumEquatableValidator extends UncheckedValidator
 {
     /**
      * Tell ruleset generator (indirectly) that this validator only accepts
-     * bool|int|string|null as allowed values.
+     * bool|int|string as allowed values.
      * @see \SimpleComplex\Validate\RuleSetFactory\RuleSetGenerator::enum()
      *
      * Overrides by prepending overriding bucket(s);
      * PHP array union(+) ignores duplicate in righthand array.
      */
     protected const PATTERN_RULES = [
-        'enum' => Type::EQUATABLE_NULL,
+        'enum' => Type::EQUATABLE,
     ]
-    + RecursiveValidator::PATTERN_RULES;
+    + UncheckedValidator::PATTERN_RULES;
 }